# apiDocs.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "OpenAPI definition",
      "version": "v0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Generated server url"
      }
    ],
    "paths": {
...
      "/api/research-contributions/{facultyId}": {
        "get": {
          "tags": [
            "research-contribution-controller"
          ],
          "operationId": "getContributionByFacultyId",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ResearchContribution"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "research-contribution-controller"
          ],
          "operationId": "updateContribution",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchContribution"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ResearchContribution"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/community-services/{facultyId}": {
        "get": {
          "tags": [
            "community-service-controller"
          ],
          "operationId": "getServiceByFacultyId",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommunityService"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "community-service-controller"
          ],
          "operationId": "updateServiceByFacultyId",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityService"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommunityService"
                  }
                }
              }
            }
          }
        }
      },
      "/api/administrative-contributions/{facultyId}": {
        "get": {
          "tags": [
            "administrative-contribution-controller"
          ],
          "operationId": "getContributionByFacultyId_1",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AdministrativeContribution"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "administrative-contribution-controller"
          ],
          "operationId": "updateContributionByFacultyId",
          "parameters": [
            {
              "name": "facultyId",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrativeContribution"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AdministrativeContribution"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/research-contributions": {
        "get": {
          "tags": [
            "research-contribution-controller"
          ],
          "operationId": "getAllContributions",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ResearchContribution"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "research-contribution-controller"
          ],
          "operationId": "createContribution",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchContribution"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ResearchContribution"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/faculties": {
        "get": {
          "tags": [
            "faculty-details-controller"
          ],
          "operationId": "getAllFaculties",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/FacultyDetails"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "faculty-details-controller"
          ],
          "operationId": "createFaculty",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacultyDetails"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/FacultyDetails"
                  }
                }
              }
            }
          }
        }
      },
      "/api/community-services": {
        "get": {
          "tags": [
            "community-service-controller"
          ],
          "operationId": "getAllServices",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CommunityService"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "community-service-controller"
          ],
          "operationId": "createService",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommunityService"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/CommunityService"
                  }
                }
              }
            }
          }
        }
      },
      "/api/administrative-contributions": {
        "get": {
          "tags": [
            "administrative-contribution-controller"
          ],
          "operationId": "getAllContributions_1",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdministrativeContribution"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "administrative-contribution-controller"
          ],
          "operationId": "createContribution_1",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrativeContribution"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AdministrativeContribution"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/faculties/{id}": {
        "get": {
          "tags": [
            "faculty-details-controller"
          ],
          "operationId": "getFacultyById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/FacultyDetails"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "faculty-details-controller"
          ],
          "operationId": "deleteFaculty",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/api/research-contributions/{id}": {
        "delete": {
          "tags": [
            "research-contribution-controller"
          ],
          "operationId": "deleteContribution",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/api/community-services/{id}": {
        "delete": {
          "tags": [
            "community-service-controller"
          ],
          "operationId": "deleteService",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/administrative-contributions/{id}": {
        "delete": {
          "tags": [
            "administrative-contribution-controller"
          ],
          "operationId": "deleteContribution_1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "FacultyDetails": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "department": {
              "type": "string"
            },
            "contactNumber": {
              "type": "string"
            },
            "joiningDate": {
              "type": "string",
              "format": "date"
            },
            "designation": {
              "type": "string"
            },
            "address": {
              "type": "string"
            }
          }
        },
...
        "ResearchContribution": {
          "type": "object",
          "properties": {
            "faculty": {
              "$ref": "#/components/schemas/FacultyDetails"
            },
            "papersPublished": {
              "type": "integer",
              "format": "int32"
            },
            "projectsCompleted": {
              "type": "integer",
              "format": "int32"
            },
            "conferencesAttended": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
...
        "CommunityService": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "faculty": {
              "$ref": "#/components/schemas/FacultyDetails"
            },
            "serviceHours": {
              "type": "integer",
              "format": "int32"
            },
            "projectsParticipated": {
              "type": "integer",
              "format": "int32"
            },
            "eventsOrganized": {
              "type": "integer",
              "format": "int32"
            },
            "foreignParticipation": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
        "AdministrativeContribution": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "faculty": {
              "$ref": "#/components/schemas/FacultyDetails"
            },
            "committeesLed": {
              "type": "integer",
              "format": "int32"
            },
            "eventsOrganized": {
              "type": "integer",
              "format": "int32"
            },
            "leadershipRoles": {
              "type": "integer",
              "format": "int32"
            }
          }
        },
...
      }
    }
  }